try to fix bug
