+++ src/components/TodoList/TodoList.vue
@@ -2,27 +2,29 @@
 <script setup lang="ts">
 import { ref } from "vue";
 
-    const appdata = {
-        data(){
-            return{
-                //inputは最初は空
-                input: "",
-                //listも最初は空
-                todo_list: []
-            }
-        },
-        methods: {
-            add_todo(){
-                //下の分でtodo_listに値を追加してく
-                this.todo_list.push(
-                    this.input
-                )
-                //ボタンを押したら入力した値が空になる
-                this.input = ""             
-               
-            }
+//幅が広い？
+
+const appdata = {
+    data(){
+        return{
+            //inputは最初は空
+            input: "",
+            //listも最初は空
+            todo_list: []
+        }
+    },
+    methods: {
+        add_todo(){
+            //下の分でtodo_listに値を追加してく
+            this.todo_list.push(
+                this.input
+            )
+            //ボタンを押したら入力した値が空になる
+            this.input = ""             
+            
         }
     }
+}
 </script>
 
 <template>
