+++ vite.config.ts
@@ -11,3 +11,6 @@ export default defineConfig({
     }
   }
 })
+//importする時「"../module"」「'@~/~'」両方で書ける
+//alias: { ... }: パスのエイリアスを指定します。@ エイリアスを、現在の
+//ファイルからの相対パス ./src にマッピングする
